[NO-ISSUE] Bug Fixes

+++ src/main/java/pl/kf/itjobsearcher/business/external/januszsoft/dto/JanuszSoftOffer.java
@@ -3,6 +3,7 @@ package pl.kf.itjobsearcher.business.external.januszsoft.dto;
 import lombok.Builder;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * Offer Class for Janusz Soft.
@@ -12,6 +13,6 @@ public record JanuszSoftOffer(
         String title,
         String description,
         String contractType,
-        BigDecimal minSalary,
-        BigDecimal maxSalary
+        BigInteger minSalary,
+        BigInteger maxSalary
 ) { }
